update program page
