+++ packages/ui/pages/account/settings.tsx
@@ -4,7 +4,7 @@ import Link from 'next/link'
 
 export default function Settings() {
   return (
-    <AppLayout>
+    <AppLayout title="User Account Settings">
       <Result
         status="404"
         title="Work in progress"
